Ref: **00/ to **24/ (#100)
Refactor **00 folders to **24/ folders to adhere to submission requirements.

Fixes #91 
WANT_APPROVAL= @remo5000  @huichun66  @DoItTomorrow 
P0= CMIW, I think the earlier we get this in, the fewer issues we'll face with merge conflicts. The alternative is to wait for every other active PR to be merged in. I don't think we'll reach a point where we have 0 active PRs. 3 is still a manageable number.