feat/qpp: Parse select-cl : declaration* ‘Select’ synonym (#146)
Implements a RDP to parse the query.
This allows the QPP to support the following grammar:
`select-cl : declaration* ‘Select’ synonym`

## State
State is a private class used only in the `QueryPreprocessor.cpp` and not exposed elsewhere. It abstracts away the logic for handling tokens and query manipulation.

## Alt design
Would have preferred to signal that there is an error in the state by returning a <state, bool> tuple in all the methods since Benedict just clarified that errors are NOT required to be thrown. This would allow for a clearer understanding of the flow of the code since nothing throws. Ultimately, it would make it easier to reason about the code. However, it is too late to be changed at the moment. I'm thinking of letting the current design be part of our tech debt as we need to finish this in less than a week.

## Testing
No unit tests as tests are covered in integration tests. (I think these are integration tests, let me know if they are not).

### Why Integration Tests?
Writing Fakes / Mocks for Lexer is too tedious.

### Not adding `/* declarationMap */` etc...
1. Tedium
2. IDE populates this tag, so it affects readability for IDE users.