feat/qpp: while pattern (#400)
Unfortunately, both the `assign` and `while` grammar rules need to be combined since they're indistinguishable.

The wiki gives a different expected outcome based on detected semantics. At the parsing level, we should still parse the tokens without semantic data. Convolution in this code is due to needing to distinguish between syntactic and semantic errors.

Upon encountering a semantic error, we should still continue to parse, but register the semantically wrong pattern clause as an invalid clause.